ci: add golangci to lefthook

+++ setup.sh
@@ -10,6 +10,18 @@ if ! [ -x "$(command -v air)" ]; then
         go install github.com/cosmtrek/air@latest
     fi
 fi
+if ! [ -x "$(command -v golangci-lint)" ]; then
+    read -p "golangci-lint is not installed. Do you want to install it via 'curl'? (y/N) " -n 1 -r
+    if [[ $REPLY =~ ^[Yy]$ ]]; then
+        # https://golangci-lint.run/usage/install/#binaries
+        # shellcheck disable=SC2046
+        curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.55.2
+    else
+        echo "Please install golangci-lint first."
+        echo "https://golangci-lint.run/usage/install/"
+        exit 1
+    fi
+fi
 if ! [ -x "$(command -v lefthook)" ]; then
     read -p "lefthook is not installed. Do you want to install it via 'go install'? (y/N) " -n 1 -r
     if [[ $REPLY =~ ^[Yy]$ ]]; then
